backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
sqlalchemy>=2.0.28
python-multipart>=0.0.9
pandas>=2.2.1
numpy>=1.26.4
python-dotenv>=1.0.1
pydantic>=2.9.0
langgraph>=0.0.48
cerebras-cloud-sdk>=1.0.0
aiofiles>=23.2.1
httpx>=0.27.0
joblib
scikit-learn